+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/mdbook-transcheck/compare/v0.2.1...Unreleased) - ReleaseDate
 
+* [Added] ignore markdown comment in original
+
 ## [v0.2.1](https://github.com/dalance/mdbook-transcheck/compare/v0.2.0...v0.2.1) - 2020-04-02
 
 * [Added] garbage line detection in code block